Time series - improvements after validation

+++ grisera_api/time_series/time_series_router.py
@@ -118,8 +118,8 @@ class TimeSeriesRouter:
     @router.get("/time_series/{time_series_id}", tags=["time series"],
                 response_model=Union[TimeSeriesOut, NotFoundByIdModel])
     async def get_time_series(self, time_series_id: int, response: Response,
-                              signal_min_value: Optional[int] = None,
-                              signal_max_value: Optional[int] = None):
+                              signal_min_value: Optional[float] = None,
+                              signal_max_value: Optional[float] = None):
         """
         Get time series by id from database with signal values.
